requirements.txt
# ============================================
# Python Dependencies for Secure Multi-Language App
# ============================================

# Flask Framework
Flask==3.0.0

# Password Hashing
bcrypt==4.1.2

# Email Support
Flask-Mail==0.9.1

# Session Management
Flask-Session==0.5.0

# Environment Variables
python-dotenv==1.0.0

# User Agent Parsing (for device/browser detection)
user-agents==2.2.0

# CORS Support
Flask-Cors==4.0.0

# Date/Time Utilities
pytz==2024.1
